fix: Moved receivers into search model
This is so that when we cancel a search, we can make sure it
no longer dispatches any more events.
