+++ src/js/models/searches/BaseSearch.js
@@ -2,6 +2,9 @@
 
 import type {DateTuple} from "../../lib/TimeWindow"
 import type {Dispatch} from "../../reducers/types"
+import {Handler} from "../../BoomClient"
+import {setBoomSearchStatus} from "../../actions/boomSearches"
+import statsReceiver from "../../receivers/statsReceiver"
 
 export default class BaseSearch {
   program: string
@@ -24,7 +27,14 @@ export default class BaseSearch {
     return this.span
   }
 
-  getReceivers(_dipatch: Dispatch) {
-    return []
+  receiveData(_handler: Handler, _dispatch: Dispatch) {}
+
+  receiveStats(handler: Handler, dispatch: Dispatch) {
+    const name = this.getName()
+    handler
+      .each(statsReceiver(name, dispatch))
+      .done(() => dispatch(setBoomSearchStatus(name, "SUCCESS")))
+      .error(() => dispatch(setBoomSearchStatus(name, "ERROR")))
+      .abort(() => dispatch(setBoomSearchStatus(name, "ABORTED")))
   }
 }
