+++ src/js/models/searches/BaseSearch.test.js
@@ -22,8 +22,4 @@ describe("BaseSearch", () => {
   test("#getSpan", () => {
     expect(search.getSpan()).toEqual([new Date(0), new Date(10)])
   })
-
-  test("#getReceivers", () => {
-    expect(search.getReceivers(jest.fn())).toEqual([])
-  })
 })
