+++ src/js/models/searches/HistogramSearch.js
@@ -1,9 +1,13 @@
 /* @flow */
 
+import throttle from "lodash/throttle"
+
 import type {Dispatch} from "../../reducers/types"
+import {Handler} from "../../BoomClient"
+import type {Payload} from "../../receivers/types"
+import {clearCountByTime, receiveCountByTime} from "../../actions/countByTime"
 import BaseSearch from "./BaseSearch"
 import countByTimeInterval from "../../lib/countByTimeInterval"
-import countByTimeReceiver from "../../receivers/countByTimeReceiver"
 
 const BOOM_INTERVALS = {
   millisecond: "ms",
@@ -23,7 +27,35 @@ export default class HistogramSearch extends BaseSearch {
     )
   }
 
-  getReceivers(dispatch: Dispatch) {
-    return [countByTimeReceiver(dispatch)]
+  receiveData(handler: Handler, dispatch: Dispatch) {
+    dispatch(clearCountByTime())
+    let descriptor
+    let tuples = []
+
+    const dispatchNow = () => {
+      if (tuples.length === 0) return
+      dispatch(receiveCountByTime({descriptor, tuples}))
+      tuples = []
+    }
+
+    const dispatchSteady = throttle(dispatchNow, 50, {leading: false})
+
+    handler
+      .channel(0, (payload: Payload) => {
+        switch (payload.type) {
+          case "SearchResult":
+            descriptor = payload.results.descriptor
+            tuples = [...tuples, ...payload.results.tuples]
+            dispatchSteady()
+            break
+          case "SearchEnd":
+            dispatchSteady.cancel()
+            dispatchNow()
+            break
+        }
+      })
+      .abort(() => {
+        dispatchSteady.cancel()
+      })
   }
 }
